+++ tests/reader_tests.rs
@@ -7,7 +7,7 @@ use clf::{pack_clf, ClfReader, MissingOpIdPolicy, PackOptions};
 /// Build a minimal .clf in memory (op_id 1 and 50, fake blobs), then open with ClfReader and get_blob.
 #[test]
 fn reader_parse_minimal_and_get_blob() {
-    let entries: Vec<(u16, Vec<u8>)> = vec![
+    let entries: Vec<(u32, Vec<u8>)> = vec![
         (1, b"blob_for_add".to_vec()),
         (50, b"blob_for_matmul".to_vec()),
     ];
@@ -53,7 +53,7 @@ fn reader_invalid_magic() {
     let mut file = tempfile::NamedTempFile::new().unwrap();
     file.write_all(b"XXXX").unwrap(); // wrong magic
     file.write_all(&[1u8]).unwrap();
-    file.write_all(&0u16.to_le_bytes()).unwrap();
+    file.write_all(&0u32.to_le_bytes()).unwrap(); // vendor len
     file.flush().unwrap();
 
     match ClfReader::open(file.path()) {
@@ -71,8 +71,8 @@ fn reader_truncated_file() {
     let mut file = tempfile::NamedTempFile::new().unwrap();
     file.write_all(&clf::CLF_MAGIC).unwrap();
     file.write_all(&[1u8]).unwrap();
-    file.write_all(&0u16.to_le_bytes()).unwrap(); // vendor len 0
-    file.write_all(&0u16.to_le_bytes()).unwrap(); // target len 0
+    file.write_all(&0u32.to_le_bytes()).unwrap(); // vendor len 0
+    file.write_all(&0u32.to_le_bytes()).unwrap(); // target len 0
     file.write_all(&[0u8]).unwrap();               // blob align 0
     // No manifest (num_entries) → read_exact will fail
     file.flush().unwrap();
@@ -86,7 +86,7 @@ fn reader_truncated_file() {
 /// build_code_section with Fail policy: missing op_id returns error.
 #[test]
 fn reader_build_code_section_fail_on_missing() {
-    let entries: Vec<(u16, Vec<u8>)> = vec![(1, b"a".to_vec())];
+    let entries: Vec<(u32, Vec<u8>)> = vec![(1, b"a".to_vec())];
     let options = PackOptions::default();
     let mut buf = Cursor::new(Vec::new());
     pack_clf(&mut buf, &entries, &options).unwrap();
@@ -96,7 +96,7 @@ fn reader_build_code_section_fail_on_missing() {
     file.flush().unwrap();
 
     let mut reader = ClfReader::open(file.path()).unwrap();
-    let op_ids = [1u16, 99]; // 99 not in CLF
+    let op_ids = [1u32, 99]; // 99 not in CLF
     let err = reader
         .build_code_section(&op_ids, MissingOpIdPolicy::Fail)
         .unwrap_err();
@@ -106,7 +106,7 @@ fn reader_build_code_section_fail_on_missing() {
 /// build_code_section with Skip policy: missing op_id skips, result non-empty.
 #[test]
 fn reader_build_code_section_skip_missing() {
-    let entries: Vec<(u16, Vec<u8>)> = vec![(1, b"a".to_vec())];
+    let entries: Vec<(u32, Vec<u8>)> = vec![(1, b"a".to_vec())];
     let options = PackOptions::default();
     let mut buf = Cursor::new(Vec::new());
     pack_clf(&mut buf, &entries, &options).unwrap();
@@ -116,7 +116,7 @@ fn reader_build_code_section_skip_missing() {
     file.flush().unwrap();
 
     let mut reader = ClfReader::open(file.path()).unwrap();
-    let op_ids = [99u16, 1]; // 99 missing, 1 present
+    let op_ids = [99u32, 1]; // 99 missing, 1 present
     let code = reader
         .build_code_section(&op_ids, MissingOpIdPolicy::Skip)
         .unwrap();
